fix: v0.3.1 local surface migration, bug fixes, and scan visibility
  - Migrate pulseengine/local/ from forward shims to full production implementations
  - Convert dashboard/, app/scan.py to reverse shims for backward compat
  - Fix dry-run writing per-asset snapshots (save=not dry_run)
  - Fix manual scan not auto-refreshing UI after completion (_scan_rerun_done)
  - Fix scan progress invisible in terminal (configure pulseengine logger at INFO, verbose=True)
  - Update test patch target to pulseengine.local.scan._save_summary
  - Update CHANGELOG, ROADMAP, CONTRIBUTING, variable_list, code_flow docs
